+++ RELEASE_NOTES.md
@@ -1,12 +1,12 @@
 ## VPinFE Release Notes
 
 ### Summary
-This is the offical release of vpinfe v2.2.1
+This is the offical release of vpinfe v2.2.3
 
 ### What's New
 None
 
 ### Fixes
-- fix "Read-only file system: '/.nicegui'" #2
+- fix stale cache after saving. #4
 
 ### Notes